Get feature importance using permutation method

Calculates feature importance using a trained model and test data. Requires the future.apply package.

get_feature_importance(
  trained_model,
  train_data,
  test_data,
  outcome_colname,
  perf_metric_function,
  perf_metric_name,
  class_probs,
  method,
  seed = NA,
  corr_thresh = 1
)

Arguments

trained_model
Trained model from caret::train().

train_data
Training data: dataframe of outcome and features.

test_data
Held out test data: dataframe of outcome and features.

outcome_colname
Column name as a string of the outcome variable (default NULL; the first column will be chosen automatically).

perf_metric_function
Function to calculate the performance metric to be used for cross-validation and test performance. Some functions are provided by caret (see caret::defaultSummary()). Defaults: binary classification = twoClassSummary, multi-class classification = multiClassSummary, regression = defaultSummary.

perf_metric_name
The column name from the output of the function provided to perf_metric_function that is to be used as the performance metric. Defaults: binary classification = "ROC", multi-class classification = "logLoss", regression = "RMSE".

class_probs
Whether to use class probabilities (TRUE for categorical outcomes, FALSE for numeric outcomes).

method
ML method. Options: c("glmnet", "rf", "rpart2", "svmRadial", "xgbTree").

glmnet: linear, logistic, or multiclass regression

rf: random forest

rpart2: decision tree

svmRadial: support vector machine

xgbTree: xgboost

seed
Random seed (default: NA). Your results will only be reproducible if you set a seed.

corr_thresh
For feature importance, group correlations above or equal to corr_thresh (range 0 to 1; default: 1).

Value

Dataframe with performance metrics for when each feature (or group of correlated features; names) is permuted (perf_metric), and differences between test performance metric and permuted performance metric (perf_metric_diff). The performance metric name (perf_metric_name) and seed (seed) are also returned.

Author

Begüm Topçuoğlu, topcuoglu.begum@gmail.com

Zena Lapp, zenalapp@umich.edu

Examples

# \donttest{
results <- run_ml(otu_small, "glmnet", kfold = 2, cv_times = 2)

#> Using 'dx' as the outcome column.

names(results$trained_model$trainingData)[1] <- "dx"
get_feature_importance(results$trained_model, results$trained_model$trainingData, results$test_data,
  "dx",
  multiClassSummary, "AUC",
  class_probs = TRUE, method = "glmnet"
)

#>    perf_metric perf_metric_diff    names method perf_metric_name seed
#> 1    0.4401579    -6.842105e-04 Otu00029 glmnet              AUC   NA
#> 2    0.4338947     5.578947e-03 Otu00028 glmnet              AUC   NA
#> 3    0.4335789     5.894737e-03 Otu00044 glmnet              AUC   NA
#> 4    0.4395526    -7.894737e-05 Otu00048 glmnet              AUC   NA
#> 5    0.4417632    -2.289474e-03 Otu00033 glmnet              AUC   NA
#> 6    0.4413158    -1.842105e-03 Otu00045 glmnet              AUC   NA
#> 7    0.4456053    -6.131579e-03 Otu00021 glmnet              AUC   NA
#> 8    0.4308684     8.605263e-03 Otu00031 glmnet              AUC   NA
#> 9    0.4428421    -3.368421e-03 Otu00038 glmnet              AUC   NA
#> 10   0.4284474     1.102632e-02 Otu00017 glmnet              AUC   NA
#> 11   0.4402105    -7.368421e-04 Otu00009 glmnet              AUC   NA
#> 12   0.4388158     6.578947e-04 Otu00039 glmnet              AUC   NA
#> 13   0.4366316     2.842105e-03 Otu00023 glmnet              AUC   NA
#> 14   0.4394737     0.000000e+00 Otu00019 glmnet              AUC   NA
#> 15   0.4425000    -3.026316e-03 Otu00026 glmnet              AUC   NA
#> 16   0.4393684     1.052632e-04 Otu00043 glmnet              AUC   NA
#> 17   0.4388421     6.315789e-04 Otu00030 glmnet              AUC   NA
#> 18   0.4394737     0.000000e+00 Otu00042 glmnet              AUC   NA
#> 19   0.4392895     1.842105e-04 Otu00032 glmnet              AUC   NA
#> 20   0.4353421     4.131579e-03 Otu00049 glmnet              AUC   NA
#> 21   0.4338158     5.657895e-03 Otu00060 glmnet              AUC   NA
#> 22   0.4017895     3.768421e-02 Otu00034 glmnet              AUC   NA
#> 23   0.4495263    -1.005263e-02 Otu00010 glmnet              AUC   NA
#> 24   0.4367105     2.763158e-03 Otu00001 glmnet              AUC   NA
#> 25   0.4394737     0.000000e+00 Otu00046 glmnet              AUC   NA
#> 26   0.4192368     2.023684e-02 Otu00015 glmnet              AUC   NA
#> 27   0.4395263    -5.263158e-05 Otu00040 glmnet              AUC   NA
#> 28   0.4531579    -1.368421e-02 Otu00024 glmnet              AUC   NA
#> 29   0.4411316    -1.657895e-03 Otu00052 glmnet              AUC   NA
#> 30   0.4396316    -1.578947e-04 Otu00016 glmnet              AUC   NA
#> 31   0.4430789    -3.605263e-03 Otu00054 glmnet              AUC   NA
#> 32   0.4326579     6.815789e-03 Otu00012 glmnet              AUC   NA
#> 33   0.4381579     1.315789e-03 Otu00007 glmnet              AUC   NA
#> 34   0.4242105     1.526316e-02 Otu00056 glmnet              AUC   NA
#> 35   0.4331579     6.315789e-03 Otu00058 glmnet              AUC   NA
#> 36   0.4396842    -2.105263e-04 Otu00022 glmnet              AUC   NA
#> 37   0.4203158     1.915789e-02 Otu00008 glmnet              AUC   NA
#> 38   0.4425789    -3.105263e-03 Otu00037 glmnet              AUC   NA
#> 39   0.4423421    -2.868421e-03 Otu00047 glmnet              AUC   NA
#> 40   0.4347105     4.763158e-03 Otu00003 glmnet              AUC   NA
#> 41   0.4423947    -2.921053e-03 Otu00011 glmnet              AUC   NA
#> 42   0.4396842    -2.105263e-04 Otu00057 glmnet              AUC   NA
#> 43   0.4321053     7.368421e-03 Otu00014 glmnet              AUC   NA
#> 44   0.4393947     7.894737e-05 Otu00006 glmnet              AUC   NA
#> 45   0.4359211     3.552632e-03 Otu00041 glmnet              AUC   NA
#> 46   0.4571579    -1.768421e-02 Otu00004 glmnet              AUC   NA
#> 47   0.4382105     1.263158e-03 Otu00050 glmnet              AUC   NA
#> 48   0.4343684     5.105263e-03 Otu00005 glmnet              AUC   NA
#> 49   0.4394737     0.000000e+00 Otu00025 glmnet              AUC   NA
#> 50   0.4416053    -2.131579e-03 Otu00036 glmnet              AUC   NA
#> 51   0.4394737     0.000000e+00 Otu00051 glmnet              AUC   NA
#> 52   0.4335526     5.921053e-03 Otu00020 glmnet              AUC   NA
#> 53   0.4353684     4.105263e-03 Otu00059 glmnet              AUC   NA
#> 54   0.4396842    -2.105263e-04 Otu00055 glmnet              AUC   NA
#> 55   0.4254474     1.402632e-02 Otu00018 glmnet              AUC   NA
#> 56   0.4391579     3.157895e-04 Otu00027 glmnet              AUC   NA
#> 57   0.4625526    -2.307895e-02 Otu00035 glmnet              AUC   NA
#> 58   0.4393684     1.052632e-04 Otu00053 glmnet              AUC   NA
#> 59   0.4348421     4.631579e-03 Otu00013 glmnet              AUC   NA
#> 60   0.4285789     1.089474e-02 Otu00002 glmnet              AUC   NA

# }