Split hyperparameters dataframe into named lists for each parameter

Using get_hyperparams_list is preferred over this function.

get_hyperparams_from_df(hyperparams_df, ml_method)

Arguments

hyperparams_df
dataframe of hyperparameters with columns param, value, and method

ml_method
machine learning method

Value

named list of lists of hyperparameters

Author

Begüm Topçuoğlu, topcuoglu.begum@gmail.com

Kelly Sovacool, sovacool@umich.edu

Examples

hparams_df <- dplyr::tibble(
  param = c("cost", "cost", "loss", "epsilon"),
  value = c(1, 0.1, "L2_primal", 0.01),
  method = rep("regLogistic", 4)
)
get_hyperparams_from_df(hparams_df, "regLogistic")

#> $cost
#> [1] "1"   "0.1"
#> 
#> $epsilon
#> [1] "0.01"
#> 
#> $loss
#> [1] "L2_primal"
#>