Select indices to partition the data into training & testing sets.

Use this function to get the row indices for the training set.

get_partition_indices(outcomes, training_frac = 0.8, groups = NULL)

Arguments

outcomes
vector of outcomes

training_frac
max fraction of data for the training set(default: 0.8)

groups
vector of groups. length must match the number of rows in the dataset. (default: NULL)

Value

Vector of row indices for the training set.

Details

If groups is NULL, uses createDataPartition. Otherwise, uses create_grouped_data_partition().

Set the seed prior to calling this function if you would like your data partitions to be reproducible (recommended).

Author

Kelly Sovacool, sovacool@umich.edu

Examples

training_inds <- get_partition_indices(otu_mini_bin$dx)
train_data <- otu_mini_bin[training_inds, ]
test_data <- otu_mini_bin[-training_inds, ]