mikropml: User-Friendly R Package for Robust Machine Learning Pipelines

mikropml implements supervised machine learning pipelines using regression, support vector machines, decision trees, random forest, or gradient-boosted trees. The main functions are preprocess_data() to process your data prior to running machine learning, and run_ml() to run machine learning.

Authors

Begüm D. Topçuoğlu (ORCID)

Zena Lapp (ORCID)

Kelly L. Sovacool (ORCID)

Evan Snitkin (ORCID)

Jenna Wiens (ORCID)

Patrick D. Schloss (ORCID)

See vignettes

Introduction

Preprocessing data

Hyperparameter tuning

Parallel processing

The mikropml paper

See also

Author

Maintainer: Kelly Sovacool sovacool@umich.edu (ORCID)

Authors:

Begüm Topçuoğlu topcuoglu.begum@gmail.com (ORCID)

Zena Lapp zenalapp@umich.edu (ORCID)

Evan Snitkin (ORCID)

Jenna Wiens (ORCID)

Patrick Schloss pschloss@umich.edu (ORCID)

Other contributors:

Nick Lesniak nlesniak@umich.edu (ORCID) [contributor]

Courtney Armour armourc@umich.edu (ORCID) [contributor]

Sarah Lucas salucas@umich.edu (ORCID) [contributor]