Randomize feature order to eliminate any position-dependent effects

Randomize feature order to eliminate any position-dependent effects

randomize_feature_order(dataset, outcome_colname)

Arguments

dataset

Dataframe with an outcome variable and other columns as features.

outcome_colname

Column name as a string of the outcome variable (default NULL; the first column will be chosen automatically).

Value

Dataset with feature order randomized.

Author

Nick Lesniak, nlesniak@umich.edu

Kelly Sovacool, sovacool@umich.edu

Examples

dat <- data.frame(
  outcome = c("1", "2", "3"),
  a = 4:6, b = 7:9, c = 10:12, d = 13:15
)
randomize_feature_order(dat, "outcome")
#>   outcome  d b a  c
#> 1       1 13 7 4 10
#> 2       2 14 8 5 11
#> 3       3 15 9 6 12