Remove missing outcome values

Remove missing outcome values

rm_missing_outcome(dataset, outcome_colname)

Arguments

dataset
dataframe with an outcome variable and other columns as features

outcome_colname
column name as a string of the outcome variable

Value

dataset with no missing outcomes

Examples

rm_missing_outcome(mikropml::otu_mini, "dx")

#>         dx Otu00001 Otu00002 Otu00003
#> 1   normal      350      268      213
#> 2   normal      568     1320       13
#> 3   normal      151      756      802
#> 4   normal      299       30     1018
#> 5   normal     1409      174        0
#> 6   normal      167      712      213
#> 7   cancer      108      120      160
#> 8   normal      347      130      131
#> 9   normal      256      995      871
#> 10  cancer     1648      144       22
#> 11  normal      843        0        0
#> 12  normal     1404      386      103
#> 13  normal      230      226      152
#> 14  normal       86      426      123
#> 15  cancer       93      540      178
#> 16  cancer     1668       20        3
#> 17  normal      381      357      315
#> 18  normal      788      558      453
#> 19  normal      282      331      250
#> 20  normal     1245       24       71
#> 21  normal     2036       21        5
#> 22  normal     1046       40       71
#> 23  cancer      322      173      419
#> 24  normal      118     1132      582
#> 25  cancer     1976      563      193
#> 26  normal     1484      327      166
#> 27  normal        7       25        8
#> 28  cancer      115       67      139
#> 29  normal      287       87      118
#> 30  normal       31      179       58
#> 31  normal      878      866      783
#> 32  normal      287      670       99
#> 33  normal     1491       74      280
#> 34  normal      587      515      162
#> 35  normal      483     1564      834
#> 36  normal      660       31       44
#> 37  normal      422      230      192
#> 38  normal      526     1172      952
#> 39  normal      166      313      182
#> 40  normal      366      317      178
#> 41  normal      641      126      170
#> 42  normal      356      369      130
#> 43  normal     1636       15        7
#> 44  normal      308       68      131
#> 45  normal      905      160       53
#> 46  normal     1777       69       64
#> 47  normal      262      235      376
#> 48  normal     1758       36       51
#> 49  normal       45      343      131
#> 50  normal     1057      463      150
#> 51  normal       62       46      122
#> 52  normal      517      560       32
#> 53  cancer     1797       92       42
#> 54  normal     1513      140        1
#> 55  cancer       99      350      229
#> 56  cancer     2072       62       54
#> 57  normal      193      210       32
#> 58  normal      387     5041        5
#> 59  normal      335     1112        2
#> 60  normal      356      129      188
#> 61  normal      454     1967        5
#> 62  normal     1063      410       64
#> 63  normal      478      158      101
#> 64  normal       64      201      146
#> 65  normal       26       91      130
#> 66  normal      222        0      757
#> 67  normal       20       42       48
#> 68  normal      136      812      229
#> 69  normal      141        0     1283
#> 70  normal      571        2      527
#> 71  normal      453       73      565
#> 72  normal      141      465      658
#> 73  cancer      296      644      320
#> 74  normal      256      165       83
#> 75  normal      152      548      158
#> 76  cancer      931       87      376
#> 77  cancer      196      547      534
#> 78  normal     1089      356     1284
#> 79  normal      130      283      121
#> 80  normal     2380        2       16
#> 81  cancer       38      208      104
#> 82  normal      361     1262      226
#> 83  normal     1334      158      128
#> 84  cancer       66       92        2
#> 85  normal       81      188     1051
#> 86  cancer      166       91      428
#> 87  normal      636      474      412
#> 88  normal      191     1110        1
#> 89  cancer      332      805      546
#> 90  normal      542        2      559
#> 91  normal       93      984      517
#> 92  cancer      329      447      257
#> 93  cancer     1583       68        0
#> 94  cancer      174      149       79
#> 95  normal      497      488      143
#> 96  cancer      401     1530     1327
#> 97  cancer      787       11        4
#> 98  cancer      229      475      286
#> 99  cancer     1697        9       20
#> 100 normal      247       22        9
#> 101 normal      477      289       94
#> 102 cancer     2474       20       12
#> 103 cancer     1216       70      103
#> 104 normal      852      298      470
#> 105 normal      250      873      112
#> 106 normal      475      569      149
#> 107 cancer      129      236        2
#> 108 cancer      388     1079      572
#> 109 cancer     1683       21       38
#> 110 cancer      223      144       63
#> 111 normal      446     1182      520
#> 112 cancer      329        1      657
#> 113 normal      332      656     1142
#> 114 cancer      100      433      165
#> 115 normal      131     2076      361
#> 116 normal      622      343      125
#> 117 cancer      203      275        0
#> 118 cancer      256      271      579
#> 119 normal      800      957     1240
#> 120 cancer      910       25       13
#> 121 normal       14      201       34
#> 122 normal      622      392        1
#> 123 cancer      151       45      105
#> 124 normal      678       38      440
#> 125 normal      865      462       62
#> 126 cancer      606     1906      363
#> 127 normal      406      354      628
#> 128 cancer      615      232       89
#> 129 normal       51      195      135
#> 130 cancer      449      435      973
#> 131 cancer     1016       98       15
#> 132 normal      380      320      300
#> 133 cancer      733      216      392
#> 134 cancer      281      549      174
#> 135 cancer      744        8       15
#> 136 normal      377     1019      452
#> 137 cancer      166      850     2336
#> 138 normal      724      224      114
#> 139 cancer       56      282       43
#> 140 normal      765      638      437
#> 141 cancer       55      556      196
#> 142 cancer       30      133      558
#> 143 cancer       50      106       75
#> 144 cancer      116      201       33
#> 145 cancer      237      293      848
#> 146 cancer      193      134       40
#> 147 cancer      401      372     1059
#> 148 cancer      250      167     1338
#> 149 cancer     1434       68       56
#> 150 normal      739     2144     1254
#> 151 cancer      249     2050      751
#> 152 cancer     1311      428      172
#> 153 normal       61      493      303
#> 154 cancer      168       36       71
#> 155 cancer      382      136       60
#> 156 cancer      401      259      169
#> 157 cancer     3540       28       26
#> 158 cancer      757      130      277
#> 159 cancer     1282       69     1331
#> 160 cancer     1401       25       14
#> 161 cancer      189       15       13
#> 162 cancer      627      649        1
#> 163 cancer      468       36       80
#> 164 normal      950      251      211
#> 165 cancer      923       31       58
#> 166 cancer        9      554      557
#> 167 cancer       18      655      496
#> 168 cancer     2671        0        2
#> 169 cancer     2723       17       27
#> 170 cancer     3360        8        5
#> 171 cancer      228      540      205
#> 172 cancer      109       89       35
#> 173 cancer      519        5        7
#> 174 cancer      420      296      131
#> 175 cancer      233      584      261
#> 176 normal      402      299     1365
#> 177 normal      275      231       54
#> 178 cancer     1927      365      218
#> 179 cancer      675       12       10
#> 180 cancer       62      579       96
#> 181 cancer       61      455      168
#> 182 cancer       71       68      242
#> 183 cancer      144       80        6
#> 184 cancer      113      749      128
#> 185 cancer      412       26        4
#> 186 cancer        4      180      234
#> 187 cancer      693      381       59
#> 188 cancer      988        7        4
#> 189 cancer       42      177      289
#> 190 cancer       98      353      146
#> 191 cancer     1139       19        0
#> 192 cancer     1321        4        1
#> 193 normal      852      449      181
#> 194 cancer      446       77      114
#> 195 cancer      266      489      146
#> 196 cancer     1806       10        7
#> 197 cancer     1434      343     1485
#> 198 normal     4170       16        0
#> 199 cancer     1284      198       60
#> 200 cancer      208      598      121

test_df <- mikropml::otu_mini
test_df[1:100, "dx"] <- NA
rm_missing_outcome(test_df, "dx")

#> Removed 100/200 (50%) of samples because of missing outcome value (NA).

#>         dx Otu00001 Otu00002 Otu00003
#> 1   normal      477      289       94
#> 2   cancer     2474       20       12
#> 3   cancer     1216       70      103
#> 4   normal      852      298      470
#> 5   normal      250      873      112
#> 6   normal      475      569      149
#> 7   cancer      129      236        2
#> 8   cancer      388     1079      572
#> 9   cancer     1683       21       38
#> 10  cancer      223      144       63
#> 11  normal      446     1182      520
#> 12  cancer      329        1      657
#> 13  normal      332      656     1142
#> 14  cancer      100      433      165
#> 15  normal      131     2076      361
#> 16  normal      622      343      125
#> 17  cancer      203      275        0
#> 18  cancer      256      271      579
#> 19  normal      800      957     1240
#> 20  cancer      910       25       13
#> 21  normal       14      201       34
#> 22  normal      622      392        1
#> 23  cancer      151       45      105
#> 24  normal      678       38      440
#> 25  normal      865      462       62
#> 26  cancer      606     1906      363
#> 27  normal      406      354      628
#> 28  cancer      615      232       89
#> 29  normal       51      195      135
#> 30  cancer      449      435      973
#> 31  cancer     1016       98       15
#> 32  normal      380      320      300
#> 33  cancer      733      216      392
#> 34  cancer      281      549      174
#> 35  cancer      744        8       15
#> 36  normal      377     1019      452
#> 37  cancer      166      850     2336
#> 38  normal      724      224      114
#> 39  cancer       56      282       43
#> 40  normal      765      638      437
#> 41  cancer       55      556      196
#> 42  cancer       30      133      558
#> 43  cancer       50      106       75
#> 44  cancer      116      201       33
#> 45  cancer      237      293      848
#> 46  cancer      193      134       40
#> 47  cancer      401      372     1059
#> 48  cancer      250      167     1338
#> 49  cancer     1434       68       56
#> 50  normal      739     2144     1254
#> 51  cancer      249     2050      751
#> 52  cancer     1311      428      172
#> 53  normal       61      493      303
#> 54  cancer      168       36       71
#> 55  cancer      382      136       60
#> 56  cancer      401      259      169
#> 57  cancer     3540       28       26
#> 58  cancer      757      130      277
#> 59  cancer     1282       69     1331
#> 60  cancer     1401       25       14
#> 61  cancer      189       15       13
#> 62  cancer      627      649        1
#> 63  cancer      468       36       80
#> 64  normal      950      251      211
#> 65  cancer      923       31       58
#> 66  cancer        9      554      557
#> 67  cancer       18      655      496
#> 68  cancer     2671        0        2
#> 69  cancer     2723       17       27
#> 70  cancer     3360        8        5
#> 71  cancer      228      540      205
#> 72  cancer      109       89       35
#> 73  cancer      519        5        7
#> 74  cancer      420      296      131
#> 75  cancer      233      584      261
#> 76  normal      402      299     1365
#> 77  normal      275      231       54
#> 78  cancer     1927      365      218
#> 79  cancer      675       12       10
#> 80  cancer       62      579       96
#> 81  cancer       61      455      168
#> 82  cancer       71       68      242
#> 83  cancer      144       80        6
#> 84  cancer      113      749      128
#> 85  cancer      412       26        4
#> 86  cancer        4      180      234
#> 87  cancer      693      381       59
#> 88  cancer      988        7        4
#> 89  cancer       42      177      289
#> 90  cancer       98      353      146
#> 91  cancer     1139       19        0
#> 92  cancer     1321        4        1
#> 93  normal      852      449      181
#> 94  cancer      446       77      114
#> 95  cancer      266      489      146
#> 96  cancer     1806       10        7
#> 97  cancer     1434      343     1485
#> 98  normal     4170       16        0
#> 99  cancer     1284      198       60
#> 100 cancer      208      598      121