Split dataset into outcome and features

Split dataset into outcome and features

split_outcome_features(dataset, outcome_colname)

Arguments

dataset
dataframe with an outcome variable and other columns as features

outcome_colname
column name as a string of the outcome variable

Value

list of length two: outcome, features

Examples

split_outcome_features(mikRopML::otu_mini,'dx')

#> $outcome
#>        dx
#> 1  normal
#> 2  normal
#> 3  normal
#> 4  normal
#> 5  normal
#> 6  normal
#> 7  cancer
#> 8  normal
#> 9  normal
#> 10 cancer
#> 11 normal
#> 12 normal
#> 13 normal
#> 14 normal
#> 15 cancer
#> 16 cancer
#> 17 normal
#> 18 normal
#> 19 normal
#> 20 normal
#> 21 normal
#> 22 normal
#> 23 cancer
#> 24 normal
#> 25 cancer
#> 26 normal
#> 27 normal
#> 28 cancer
#> 29 normal
#> 30 normal
#> 31 normal
#> 32 normal
#> 33 normal
#> 34 normal
#> 35 normal
#> 36 normal
#> 37 normal
#> 38 normal
#> 39 normal
#> 40 normal
#> 41 normal
#> 42 normal
#> 43 normal
#> 44 normal
#> 45 normal
#> 46 normal
#> 47 normal
#> 48 normal
#> 49 normal
#> 50 normal
#> 51 normal
#> 52 normal
#> 53 cancer
#> 54 normal
#> 55 cancer
#> 56 cancer
#> 57 normal
#> 58 normal
#> 59 normal
#> 60 normal
#> 
#> $features
#>    Otu00001 Otu00002 Otu00003
#> 1       350      268      213
#> 2       568     1320       13
#> 3       151      756      802
#> 4       299       30     1018
#> 5      1409      174        0
#> 6       167      712      213
#> 7       108      120      160
#> 8       347      130      131
#> 9       256      995      871
#> 10     1648      144       22
#> 11      843        0        0
#> 12     1404      386      103
#> 13      230      226      152
#> 14       86      426      123
#> 15       93      540      178
#> 16     1668       20        3
#> 17      381      357      315
#> 18      788      558      453
#> 19      282      331      250
#> 20     1245       24       71
#> 21     2036       21        5
#> 22     1046       40       71
#> 23      322      173      419
#> 24      118     1132      582
#> 25     1976      563      193
#> 26     1484      327      166
#> 27        7       25        8
#> 28      115       67      139
#> 29      287       87      118
#> 30       31      179       58
#> 31      878      866      783
#> 32      287      670       99
#> 33     1491       74      280
#> 34      587      515      162
#> 35      483     1564      834
#> 36      660       31       44
#> 37      422      230      192
#> 38      526     1172      952
#> 39      166      313      182
#> 40      366      317      178
#> 41      641      126      170
#> 42      356      369      130
#> 43     1636       15        7
#> 44      308       68      131
#> 45      905      160       53
#> 46     1777       69       64
#> 47      262      235      376
#> 48     1758       36       51
#> 49       45      343      131
#> 50     1057      463      150
#> 51       62       46      122
#> 52      517      560       32
#> 53     1797       92       42
#> 54     1513      140        1
#> 55       99      350      229
#> 56     2072       62       54
#> 57      193      210       32
#> 58      387     5041        5
#> 59      335     1112        2
#> 60      356      129      188
#>