Preprocess data prior to running machine learning

Preprocess data prior to running machine learning

preprocess_data(
  dataset,
  outcome_colname,
  method = c("center", "scale"),
  rm_nzv = TRUE
)

Arguments

dataset
dataframe with an outcome variable and other columns as features

outcome_colname
column name as a string of the outcome variable

method
methods to preprocess the data, described in caret::preProcess (defaut: c("center","scale"), use NULL for no normalization)

rm_nzv
whether to remove variables with near-zero variance (default: TRUE)

Value

preprocessed data

Examples

preprocess_data(mikRopML::otu_small,'dx')

#> # A tibble: 60 x 61
#>    dx    Otu00001 Otu00002 Otu00003 Otu00004 Otu00005 Otu00006 Otu00007 Otu00008
#>    <fct>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1 norm…   -0.555   -0.223   0.0166  -0.570    0.0199   -0.327   -0.605   0.0818
#>  2 norm…   -0.204    1.27   -0.769    0.0182   1.77     -0.669   -0.679  -0.0286
#>  3 norm…   -0.875    0.469   2.33     0.548   -0.218    -0.216   -0.649  -0.148 
#>  4 norm…   -0.637   -0.560   3.18    -0.572   -0.670    -0.680   -0.589  -0.564 
#>  5 norm…    1.15    -0.356  -0.820   -0.566   -0.757     2.12     0.152  -0.891 
#>  6 norm…   -0.849    0.406   0.0166  -0.564    0.487     0.494   -0.374   0.171 
#>  7 canc…   -0.944   -0.432  -0.191    0.131   -0.651     1.67    -0.753   0.282 
#>  8 norm…   -0.559   -0.418  -0.305   -0.537   -0.357    -0.237    0.336   0.146 
#>  9 norm…   -0.706    0.807   2.60    -0.346   -0.229    -0.578   -0.840   0.647 
#> 10 canc…    1.53    -0.398  -0.733   -0.0704  -0.704     0.397   -0.408   2.70  
#> # … with 50 more rows, and 52 more variables: Otu00009 <dbl>, Otu00010 <dbl>,
#> #   Otu00011 <dbl>, Otu00012 <dbl>, Otu00013 <dbl>, Otu00014 <dbl>,
#> #   Otu00015 <dbl>, Otu00016 <dbl>, Otu00017 <dbl>, Otu00018 <dbl>,
#> #   Otu00019 <dbl>, Otu00020 <dbl>, Otu00021 <dbl>, Otu00022 <dbl>,
#> #   Otu00023 <dbl>, Otu00024 <dbl>, Otu00025 <dbl>, Otu00026 <dbl>,
#> #   Otu00027 <dbl>, Otu00028 <dbl>, Otu00029 <dbl>, Otu00030 <dbl>,
#> #   Otu00031 <dbl>, Otu00032 <dbl>, Otu00033 <dbl>, Otu00034 <dbl>,
#> #   Otu00035 <dbl>, Otu00036 <dbl>, Otu00037 <dbl>, Otu00038 <dbl>,
#> #   Otu00039 <dbl>, Otu00040 <dbl>, Otu00041 <dbl>, Otu00042 <dbl>,
#> #   Otu00043 <dbl>, Otu00044 <dbl>, Otu00045 <dbl>, Otu00046 <dbl>,
#> #   Otu00047 <dbl>, Otu00048 <dbl>, Otu00049 <dbl>, Otu00050 <dbl>,
#> #   Otu00051 <dbl>, Otu00052 <dbl>, Otu00053 <dbl>, Otu00054 <dbl>,
#> #   Otu00055 <dbl>, Otu00056 <dbl>, Otu00057 <dbl>, Otu00058 <dbl>,
#> #   Otu00059 <dbl>, Otu00060 <dbl>