Run the machine learning pipeline

Run the machine learning pipeline

run_ml(
  dataset,
  method,
  outcome_colname = NULL,
  outcome_value = NULL,
  hyperparameters = NULL,
  find_feature_importance = FALSE,
  kfold = 5,
  cv_times = 100,
  training_frac = 0.8,
  group = NULL,
  corr_thresh = 1,
  seed = NA,
  ncores = NA
)

Arguments

dataset
dataframe with an outcome variable and other columns as features

method
ML method ("regLogistic", "rf", "rpart2", "svmRadial", "xgbTree")

outcome_colname
column name as a string of the outcome variable

outcome_value
outcome value of interest as a string

hyperparameters
dataframe of hyperparameters (default: NULL). if NULL given, they will be chosen automatically.

find_feature_importance
run permutation imporance (default: FALSE)

kfold
fold number for k-fold cross-validation (default: 5)

cv_times
Number of partitions to create

training_frac
fraction size of data for training (default: 0.8)

group
vector of groups to keep together when splitting the data into train and test sets, and for cross-validation; length matches the number of rows in the dataset (default: no groups)

corr_thresh
for feature importance, group correlations above or equal to corr_thresh (default: 1)

seed
random seed (default: NA)

ncores
number of cores for parallel processing (default: NA). parallel and doParallel packages are needed for ncores > 1

Value

named list with results

Author

Begüm Topçuoğlu, topcuoglu.begum@gmail.com

Zena Lapp, zenalapp@umich.edu

Kelly Sovacool, sovacool@umich.edu

Examples

if (FALSE) {
run_ml(otu_large, "regLogistic")
run_ml(otu_mini, "regLogistic",
  kfold = 2,
  find_feature_importance = TRUE
)
}