Run the machine learning pipeline

TODO: more details

run_ml(
  dataset,
  method,
  outcome_colname = NULL,
  hyperparameters = NULL,
  find_feature_importance = FALSE,
  kfold = 5,
  cv_times = 100,
  training_frac = 0.8,
  perf_metric_function = NULL,
  perf_metric_name = NULL,
  groups = NULL,
  corr_thresh = 1,
  ntree = 1000,
  seed = NA
)

Arguments

dataset
Dataframe with an outcome variable and other columns as features.

method
ML method. Options: c("glmnet", "rf", "rpart2", "svmRadial", "xgbTree")

outcome_colname
Column name as a string of the outcome variable (default NULL; will be chosen automatically).

hyperparameters
Dataframe of hyperparameters (default NULL; will be chosen automatically).

find_feature_importance
Run permutation imporance (default: FALSE). This is recommended, but it is resource-intensive.

kfold
Fold number for k-fold cross-validation (default: 5).

cv_times
Number of partitions to create (default: 100).

training_frac
Fraction of data for training set (default: 0.8). The remaining data will be used in the testing set.

perf_metric_function
Function to calculate the performance metric to be used for cross-validation and test performance. Some functions are provided by caret (see defaultSummary). Defaults: binary classification = twoClassSummary, multi-class classification = multiClassSummary, regression = defaultSummary.

perf_metric_name
The column name from the output of the function provided to perf_metric_function that is to be used as the performance metric. Defaults: binary classification = "ROC", multi-class classification = "logLoss", regression = "RMSE".

groups
Vector of groups to keep together when splitting the data into train and test sets, and for cross-validation; length matches the number of rows in the dataset (default: no groups).

corr_thresh
For feature importance, group correlations above or equal to corr_thresh (default: 1).

ntree
For random forest, how many trees to use (default: 1000). Note that caret doesn't allow this parameter to be tuned.

seed
Random seed (default: NA). Your results will be reproducible if you set a seed.

Value

named list with results

Author

Begüm Topçuoğlu, topcuoglu.begum@gmail.com

Zena Lapp, zenalapp@umich.edu

Kelly Sovacool, sovacool@umich.edu

Examples

if (FALSE) {
run_ml(otu_small, "glmnet",
  kfold = 2,
  cv_times = 2,
  seed = 2019
)
}