Split dataset into outcome and features

Split dataset into outcome and features

split_outcome_features(dataset, outcome_colname)

Arguments

dataset
dataframe with an outcome variable and other columns as features

outcome_colname
column name as a string of the outcome variable

Value

list of length two: outcome, features

Examples

split_outcome_features(mikropml::otu_mini, "dx")

#> $outcome
#>         dx
#> 1   normal
#> 2   normal
#> 3   normal
#> 4   normal
#> 5   normal
#> 6   normal
#> 7   cancer
#> 8   normal
#> 9   normal
#> 10  cancer
#> 11  normal
#> 12  normal
#> 13  normal
#> 14  normal
#> 15  cancer
#> 16  cancer
#> 17  normal
#> 18  normal
#> 19  normal
#> 20  normal
#> 21  normal
#> 22  normal
#> 23  cancer
#> 24  normal
#> 25  cancer
#> 26  normal
#> 27  normal
#> 28  cancer
#> 29  normal
#> 30  normal
#> 31  normal
#> 32  normal
#> 33  normal
#> 34  normal
#> 35  normal
#> 36  normal
#> 37  normal
#> 38  normal
#> 39  normal
#> 40  normal
#> 41  normal
#> 42  normal
#> 43  normal
#> 44  normal
#> 45  normal
#> 46  normal
#> 47  normal
#> 48  normal
#> 49  normal
#> 50  normal
#> 51  normal
#> 52  normal
#> 53  cancer
#> 54  normal
#> 55  cancer
#> 56  cancer
#> 57  normal
#> 58  normal
#> 59  normal
#> 60  normal
#> 61  normal
#> 62  normal
#> 63  normal
#> 64  normal
#> 65  normal
#> 66  normal
#> 67  normal
#> 68  normal
#> 69  normal
#> 70  normal
#> 71  normal
#> 72  normal
#> 73  cancer
#> 74  normal
#> 75  normal
#> 76  cancer
#> 77  cancer
#> 78  normal
#> 79  normal
#> 80  normal
#> 81  cancer
#> 82  normal
#> 83  normal
#> 84  cancer
#> 85  normal
#> 86  cancer
#> 87  normal
#> 88  normal
#> 89  cancer
#> 90  normal
#> 91  normal
#> 92  cancer
#> 93  cancer
#> 94  cancer
#> 95  normal
#> 96  cancer
#> 97  cancer
#> 98  cancer
#> 99  cancer
#> 100 normal
#> 101 normal
#> 102 cancer
#> 103 cancer
#> 104 normal
#> 105 normal
#> 106 normal
#> 107 cancer
#> 108 cancer
#> 109 cancer
#> 110 cancer
#> 111 normal
#> 112 cancer
#> 113 normal
#> 114 cancer
#> 115 normal
#> 116 normal
#> 117 cancer
#> 118 cancer
#> 119 normal
#> 120 cancer
#> 121 normal
#> 122 normal
#> 123 cancer
#> 124 normal
#> 125 normal
#> 126 cancer
#> 127 normal
#> 128 cancer
#> 129 normal
#> 130 cancer
#> 131 cancer
#> 132 normal
#> 133 cancer
#> 134 cancer
#> 135 cancer
#> 136 normal
#> 137 cancer
#> 138 normal
#> 139 cancer
#> 140 normal
#> 141 cancer
#> 142 cancer
#> 143 cancer
#> 144 cancer
#> 145 cancer
#> 146 cancer
#> 147 cancer
#> 148 cancer
#> 149 cancer
#> 150 normal
#> 151 cancer
#> 152 cancer
#> 153 normal
#> 154 cancer
#> 155 cancer
#> 156 cancer
#> 157 cancer
#> 158 cancer
#> 159 cancer
#> 160 cancer
#> 161 cancer
#> 162 cancer
#> 163 cancer
#> 164 normal
#> 165 cancer
#> 166 cancer
#> 167 cancer
#> 168 cancer
#> 169 cancer
#> 170 cancer
#> 171 cancer
#> 172 cancer
#> 173 cancer
#> 174 cancer
#> 175 cancer
#> 176 normal
#> 177 normal
#> 178 cancer
#> 179 cancer
#> 180 cancer
#> 181 cancer
#> 182 cancer
#> 183 cancer
#> 184 cancer
#> 185 cancer
#> 186 cancer
#> 187 cancer
#> 188 cancer
#> 189 cancer
#> 190 cancer
#> 191 cancer
#> 192 cancer
#> 193 normal
#> 194 cancer
#> 195 cancer
#> 196 cancer
#> 197 cancer
#> 198 normal
#> 199 cancer
#> 200 cancer
#> 
#> $features
#>     Otu00001 Otu00002 Otu00003
#> 1        350      268      213
#> 2        568     1320       13
#> 3        151      756      802
#> 4        299       30     1018
#> 5       1409      174        0
#> 6        167      712      213
#> 7        108      120      160
#> 8        347      130      131
#> 9        256      995      871
#> 10      1648      144       22
#> 11       843        0        0
#> 12      1404      386      103
#> 13       230      226      152
#> 14        86      426      123
#> 15        93      540      178
#> 16      1668       20        3
#> 17       381      357      315
#> 18       788      558      453
#> 19       282      331      250
#> 20      1245       24       71
#> 21      2036       21        5
#> 22      1046       40       71
#> 23       322      173      419
#> 24       118     1132      582
#> 25      1976      563      193
#> 26      1484      327      166
#> 27         7       25        8
#> 28       115       67      139
#> 29       287       87      118
#> 30        31      179       58
#> 31       878      866      783
#> 32       287      670       99
#> 33      1491       74      280
#> 34       587      515      162
#> 35       483     1564      834
#> 36       660       31       44
#> 37       422      230      192
#> 38       526     1172      952
#> 39       166      313      182
#> 40       366      317      178
#> 41       641      126      170
#> 42       356      369      130
#> 43      1636       15        7
#> 44       308       68      131
#> 45       905      160       53
#> 46      1777       69       64
#> 47       262      235      376
#> 48      1758       36       51
#> 49        45      343      131
#> 50      1057      463      150
#> 51        62       46      122
#> 52       517      560       32
#> 53      1797       92       42
#> 54      1513      140        1
#> 55        99      350      229
#> 56      2072       62       54
#> 57       193      210       32
#> 58       387     5041        5
#> 59       335     1112        2
#> 60       356      129      188
#> 61       454     1967        5
#> 62      1063      410       64
#> 63       478      158      101
#> 64        64      201      146
#> 65        26       91      130
#> 66       222        0      757
#> 67        20       42       48
#> 68       136      812      229
#> 69       141        0     1283
#> 70       571        2      527
#> 71       453       73      565
#> 72       141      465      658
#> 73       296      644      320
#> 74       256      165       83
#> 75       152      548      158
#> 76       931       87      376
#> 77       196      547      534
#> 78      1089      356     1284
#> 79       130      283      121
#> 80      2380        2       16
#> 81        38      208      104
#> 82       361     1262      226
#> 83      1334      158      128
#> 84        66       92        2
#> 85        81      188     1051
#> 86       166       91      428
#> 87       636      474      412
#> 88       191     1110        1
#> 89       332      805      546
#> 90       542        2      559
#> 91        93      984      517
#> 92       329      447      257
#> 93      1583       68        0
#> 94       174      149       79
#> 95       497      488      143
#> 96       401     1530     1327
#> 97       787       11        4
#> 98       229      475      286
#> 99      1697        9       20
#> 100      247       22        9
#> 101      477      289       94
#> 102     2474       20       12
#> 103     1216       70      103
#> 104      852      298      470
#> 105      250      873      112
#> 106      475      569      149
#> 107      129      236        2
#> 108      388     1079      572
#> 109     1683       21       38
#> 110      223      144       63
#> 111      446     1182      520
#> 112      329        1      657
#> 113      332      656     1142
#> 114      100      433      165
#> 115      131     2076      361
#> 116      622      343      125
#> 117      203      275        0
#> 118      256      271      579
#> 119      800      957     1240
#> 120      910       25       13
#> 121       14      201       34
#> 122      622      392        1
#> 123      151       45      105
#> 124      678       38      440
#> 125      865      462       62
#> 126      606     1906      363
#> 127      406      354      628
#> 128      615      232       89
#> 129       51      195      135
#> 130      449      435      973
#> 131     1016       98       15
#> 132      380      320      300
#> 133      733      216      392
#> 134      281      549      174
#> 135      744        8       15
#> 136      377     1019      452
#> 137      166      850     2336
#> 138      724      224      114
#> 139       56      282       43
#> 140      765      638      437
#> 141       55      556      196
#> 142       30      133      558
#> 143       50      106       75
#> 144      116      201       33
#> 145      237      293      848
#> 146      193      134       40
#> 147      401      372     1059
#> 148      250      167     1338
#> 149     1434       68       56
#> 150      739     2144     1254
#> 151      249     2050      751
#> 152     1311      428      172
#> 153       61      493      303
#> 154      168       36       71
#> 155      382      136       60
#> 156      401      259      169
#> 157     3540       28       26
#> 158      757      130      277
#> 159     1282       69     1331
#> 160     1401       25       14
#> 161      189       15       13
#> 162      627      649        1
#> 163      468       36       80
#> 164      950      251      211
#> 165      923       31       58
#> 166        9      554      557
#> 167       18      655      496
#> 168     2671        0        2
#> 169     2723       17       27
#> 170     3360        8        5
#> 171      228      540      205
#> 172      109       89       35
#> 173      519        5        7
#> 174      420      296      131
#> 175      233      584      261
#> 176      402      299     1365
#> 177      275      231       54
#> 178     1927      365      218
#> 179      675       12       10
#> 180       62      579       96
#> 181       61      455      168
#> 182       71       68      242
#> 183      144       80        6
#> 184      113      749      128
#> 185      412       26        4
#> 186        4      180      234
#> 187      693      381       59
#> 188      988        7        4
#> 189       42      177      289
#> 190       98      353      146
#> 191     1139       19        0
#> 192     1321        4        1
#> 193      852      449      181
#> 194      446       77      114
#> 195      266      489      146
#> 196     1806       10        7
#> 197     1434      343     1485
#> 198     4170       16        0
#> 199     1284      198       60
#> 200      208      598      121
#>